libpod: Move functions related to /etc bind mounts to container_internal_common.go
[NO NEW TESTS NEEDED]
